Do not delete `inline` keyword if not defined
GCC (and probably other compilers) supports the 'inline' keyword, but
does not define it (in the preprocessor directive sense). The lines
removed in this commit delete the `inline` keyword from the source
files in the preprocessing stage of a GCC compile.

Also, the MinGW (and Mingw-w64) w32api headers use `inline` in the
definition of some of the functions (at least in `stdlib.h`). Deleting
the `inline` keyword from them results in "multiple definition of
<function name>" errors at link-time.

This commit partially reverts changes made in commit
5991310f87724aaa1d47850baf75c2512c269cef ("Fix jbig2dec compilation
when inline is not defined").